% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_neighbors.R
\name{annot_neighbors}
\alias{annot_neighbors}
\title{Compute adjacency matrix of surface parcellation (annot) regions.}
\usage{
annot_neighbors(annot_min, surface, empty_rename = "_")
}
\arguments{
\item{annot_min}{a full \code{fs.annot} instance, or a minimal annot, i.e., only the 'label_names' field of the annot. Only a single one, not a \code{\link{hemilist}}.}

\item{surface}{\code{fs.surface} instance, the brain mesh (for a single hemisphere).}

\item{empty_rename}{optional character string, the name to use for unnamed regions (which have an empty string as region name).}
}
\value{
named integer matrix of regions, expressing whether they are direct neighbors (value \code{1L}) or not (value \code{0L}).
}
\description{
Given a brain surface and a parcellation into regions, find out which regions are adjacent to each other. A region \code{i} is adjacent to another region \code{j} if an edge connects any vertex of
\code{i} with any vertex of \code{i}.
}
\note{
Regions from different hemispheres cannot be direct neighbors.

This works on a single hemisphere. We should create a wrapper that works on brainparc instances.
}
\keyword{internal}
