% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{brainparc_fs}
\alias{brainparc_fs}
\title{Create a brain parcellation from a FreeSurfer annotation.}
\usage{
brainparc_fs(
  subjects_dir,
  subject_id,
  surface = "white",
  atlas = c("aparc", "aparc.a2009s")
)
}
\arguments{
\item{subjects_dir}{string. The FreeSurfer SUBJECTS_DIR, i.e., a directory containing the data for all your subjects, each in a subdir named after the subject identifier.}

\item{subject_id}{string. The subject identifier}

\item{surface}{string. The surface name. E.g., "white", or "pial". Used to construct the name of the surface file to be loaded.}

\item{atlas}{string. The atlas name. E.g., "aparc", "aparc.2009s", or "aparc.DKTatlas". Used to construct the name of the annotation file to be loaded.}
}
\description{
Create a brain parcellation from a FreeSurfer annotation.
}
