% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_neighbors.R
\name{brainparc_neighbors}
\alias{brainparc_neighbors}
\title{Compute adjacency of surface parcellation (annot) regions.}
\usage{
brainparc_neighbors(bp)
}
\arguments{
\item{bp}{a brainparc instance}
}
\value{
hemilist of named integer matrix of regions, expressing whether they are direct neighbors (value \code{1L}) or not (value \code{0L}).
}
\description{
Given a brain surface and a parcellation into regions, find out which regions are adjacent to each other. A region \code{i} is adjacent to another region \code{j} if an edge connects any vertex of
\code{i} with any vertex of \code{i}. Works with brainparcellations.
}
