% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_regions.R
\name{center_of_mass}
\alias{center_of_mass}
\title{Compute the center of mass of the given points.}
\usage{
center_of_mass(coords, weights = NULL)
}
\arguments{
\item{coords}{numeric \code{nx3} matrix if point x,y,z coordinates. A single coordinate can be passed as a vector of length 3, and will be returned as is.}

\item{weights}{numerical vector of length \code{n}, weights for the points in \code{coords}. Assumed to be all \code{1.0} if omitted.}
}
\value{
vector of length 3, the center of mass.
}
\description{
Compute the center of mass of the given points.
}
\keyword{internal}
