% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_annot.R
\name{cluster_overlapping_regions}
\alias{cluster_overlapping_regions}
\title{Given vertex indices defining a cluster, find all atlas regions the cluster overlaps with.}
\usage{
cluster_overlapping_regions(annot_min, cluster_vertices)
}
\arguments{
\item{annot_min}{a full \code{fs.annot} instance, or a minimal annot, i.e., only the 'label_names' field of the annot. Only a single one, not a \code{\link{hemilist}}.}

\item{cluster_vertices}{integer vector, the vertices defining the cluster (technically they do not need to form a cluster or be connected). Must not be empty.}
}
\value{
data.frame with columns 'region': the region name, 'num_shared_vertices': the number of cluster vertices which are in the region, and 'percent_shared_vertices': the percent of cluster vertices which are in the region, 'cluster_percent_of_region': how much of the region area is filled by the cluster vertices, in percent. The data.frame rows are ordered descending by 'percent_shared_vertices'.
}
\description{
Given vertex indices defining a cluster, find all atlas regions the cluster overlaps with.
}
\keyword{internal}
