% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loc.R
\name{cluster_region_overlap}
\alias{cluster_region_overlap}
\title{Compute overlapping atlas regions for clusters.}
\usage{
cluster_region_overlap(
  clusterinfo,
  silent = getOption("brainloc.silent", default = FALSE)
)
}
\arguments{
\item{clusterinfo}{a clusterinfo instance, see the \code{clusterinfo} function to get one.}

\item{silent}{whether to suppress printing messages to stdout.}
}
\value{
a data.frame with cluster overlap information.
}
\description{
For each cluster in the clusterinfo, compute 1) which atlas regions it overlaps with, and to what percentage the cluster is contained in each region. And, 2), how much of each of the overlapping regions are covered by the cluster. The atlases or surface parcellations to compare the clusters to are take from the \code{clusterinfo$brainparc} field.
}
