% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterinfo.R
\name{clusterinfo_from_thresholded_overlay}
\alias{clusterinfo_from_thresholded_overlay}
\title{Create clusterinfo instance from thresholded per-vertex data overlays.}
\usage{
clusterinfo_from_thresholded_overlay(
  lh_threshmap,
  rh_threshmap,
  value_thresholded = 0,
  template_subject = "fsaverage",
  subjects_dir = file.path(getOption("brainloc.fs_home", default =
    Sys.getenv("FREESURFER_HOME")), "subjects")
)
}
\arguments{
\item{lh_threshmap}{double vector, the stats map. Typically a thresholded t-value map. Must have one value per vertex. The value assigned to vertices that have been removed by the thresholding can be set with parameter 'value_thresholded'. If a character string, the parameter will be interpreted as file path and loaded with \code{freesurferformats::read.fs.morph}.}

\item{rh_threshmap}{like \code{lh_threshmap}, but for the right hemisphere.}

\item{value_thresholded}{scalar double, the data value of thresholded vertices in the \code{lh_threshmap} and \code{rh_threshmap}.}

\item{template_subject}{character string, the template subject name. Typically 'fsaverage' or 'fsaverage6'. Must be a subject in MNI305 space (i.e., the standard FreeSurfer template space). **Important:** The overlay and statmap data passed to this function must be in the same space as the template_subject (usually MNI305 / fsaverage space). Passing data from a different space (e.g., native subject space) will produce silently incorrect coordinate transforms and atlas region assignments.}

\item{subjects_dir}{character string, file system path to a directory containing the recon-all data for the template_subject. Used to load surfaces and annotations to identify cluster coordinates and atlas regions.}
}
\value{
a \code{clusterinfo} instance.
}
\description{
Create clusterinfo instance from thresholded per-vertex data overlays.
}
\note{
This can only work if there are no clusters which touch each other. I think this can never happen, but we should double-check.
}
\seealso{
\code{\link{clusterinfo}} If you do have a separate t-map and a cluster overlay map instead of only a thresholded t-map.
}
