% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusterdata.R
\name{clusters_to_annot}
\alias{clusters_to_annot}
\title{Create a hemilist of fs.annot instances from the given cluster overlay.}
\usage{
clusters_to_annot(clusterdata, background_code = 0L, hemi = NULL)
}
\arguments{
\item{clusterdata}{an integer vector of cluster data: a vector that assigns each vertex to an integer class, all vertices belonging to a cluster share the same integer assignment. Non-cluster vertices are assigned the background class, see parameter 'background_code'.}

\item{background_code}{scalar integer, the code in the data that should be interpreted as background or 'not part of any cluster'.}

\item{hemi}{character string, ignore unless clusterdata is a vector. In that case, it must be 'lh' or 'rh'. Used to name clusters, as a prefix.}
}
\description{
Create a hemilist of fs.annot instances from the given cluster overlay.
}
\keyword{internal}
