% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coord_fsaverage_to_MNI152}
\alias{coord_fsaverage_to_MNI152}
\title{Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.}
\usage{
coord_fsaverage_to_MNI152(vertex_coords)
}
\arguments{
\item{vertex_coords}{nx3 matrix of coordinates, e.g., typically from fsaverage surface vertices.}
}
\value{
nx3 numerical matrix if MNI152 coords.
}
\description{
Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.
}
\note{
One can check that the results are okay by clicking a vertex in FreeView, using the displayed MNI305 coordinates as input to this function, and looking up the reported MNI152 coordinates at \code{https://bioimagesuiteweb.github.io/webapp/mni2tal.html}.
}
