% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coord_fssurface_to_talairach}
\alias{coord_fssurface_to_talairach}
\title{Transform FreeSurfer surface space coordinates into Talairach space.}
\usage{
coord_fssurface_to_talairach(subjects_dir, subject_id, surface_coords)
}
\arguments{
\item{subjects_dir}{character string, the path to the recon-all SUBJECTS_DIR containing the data of all subjects of your study.}

\item{subject_id}{character string, the subject identifier (i.e., the sub directory name under the \code{subjects_dir}).}

\item{surface_coords}{nx3 numerical matrix of surface coordinates (vertex positions in FreeSurfer surface space).}
}
\value{
nx3 numerical matrix of Talairach space coordinates.
}
\description{
This uses the affine transform in \code{subject/mri/transforms/talairach.xfm}, be sure to know about the limitations of this approach. See the FreeSurfer documentation on talairach for details.
}
\examples{
\dontrun{
# will lead to identity transform, so no change in coords.
sjd = fsbrain::fsaverage.path();
sj = "fsaverage";
lh_white = fsbrain::subject.surface(sjd, sj, hemi="lh");
coord_fssurface_to_talairach(sjd, sj, lh_white$vertices[1:4,]);
}

}
