% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coord_MNI305_info}
\alias{coord_MNI305_info}
\title{Get info on MNI305 surface coordinate in various coordinate systems.}
\usage{
coord_MNI305_info(coords_mni305, ...)
}
\arguments{
\item{coords_mni305}{nx3 matrix of coordinates, typically from fsaverage surface vertices.}

\item{...}{Passed on to \code{\link{coord_MNI305_to_MNI152}}.}
}
\value{
named list with coordinate information
}
\description{
Get info on MNI305 surface coordinate in various coordinate systems.
}
