% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coords_fsaverage_to_MNI152}
\alias{coords_fsaverage_to_MNI152}
\title{Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.}
\usage{
coords_fsaverage_to_MNI152(vertex_coords)
}
\arguments{
\item{vertex_coords}{nx3 matrix of coordinates, e.g., typically from fsaverage surface vertices.}
}
\value{
nx3 numerical matrix if MNI152 coords.
}
\description{
Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.
}
