% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optdata.R
\name{download_talairach}
\alias{download_talairach}
\title{Download the talairach volume and labels from talairach.org.}
\usage{
download_talairach(accept_talairach_usage = TRUE)
}
\value{
Named list. The list has entries: "available": vector of strings. The names of the files that are available in the local file cache. You can access them using get_optional_data_file(). "missing": vector of strings. The names of the files that this function was unable to retrieve.
}
\description{
Download the talairach volume \code{talairach.nii} and labels \code{labels.txt} from talairach.org.
}
\note{
This function requires and internet connection. Files will only be downloaded if they are not already available on the local computer.
}
\keyword{internal}
