% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fs.surface.to.tmesh3d}
\alias{fs.surface.to.tmesh3d}
\title{Convert fs.surface to tmesh3d using freesurferformats of fsbrain.}
\usage{
fs.surface.to.tmesh3d(surface)
}
\arguments{
\item{surface}{an fs.surface instance.}
}
\value{
\code{tmesh} instance, as used in the \code{rgl} package.
}
\description{
Convert fs.surface to tmesh3d using freesurferformats of fsbrain.
}
\keyword{internal}
