% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fs_home}
\alias{fs_home}
\title{Return FreeSurfer path.}
\usage{
fs_home()
}
\value{
the FreeSurfer path, typically what the environment variable `FREESURFER_HOME` points to.
}
\description{
Return FreeSurfer path.
}
\note{
This function will stop (i.e., raise an error) if the directory cannot be found. It calls \code{\link{find.freesurferhome}} internally, see there for more details.
}
\seealso{
Other FreeSurfer helper functions:
\code{\link[=find.freesurferhome]{find.freesurferhome()}},
\code{\link[=get_subjects_dir]{get_subjects_dir()}},
\code{\link[=has_fs]{has_fs()}}
}
\concept{FreeSurfer helper functions}
