% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusterdata.R
\name{get_cluster_location_details}
\alias{get_cluster_location_details}
\title{Get details on cluster location, required brainparc.}
\usage{
get_cluster_location_details(clusters)
}
\arguments{
\item{clusters}{a clusterinfo instance, see the \code{clusterinfo} function on how to get one.}
}
\value{
a new version of the input data.frame, with additional columns appended.
}
\description{
Get details on cluster location, required brainparc.
}
