% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{get_minannot}
\alias{get_minannot}
\title{Extract minannot from fs.annot instance.}
\usage{
get_minannot(fs.annot)
}
\arguments{
\item{fs.annot}{a \code{freesurferformats::fs.annot} instance, or a \code{minannot} instance, which will be returned as-is. Alternatively hemilist of the latter.}
}
\value{
a \code{minannot} instance, i.e., only the 'label_names' field of the \code{fs.annot} instance.
}
\description{
Extract minannot from fs.annot instance.
}
\keyword{internal}
