% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_subjects_dir}
\alias{get_subjects_dir}
\title{Find subjects_dir or stop.}
\usage{
get_subjects_dir(
  mustWork = TRUE,
  allow_download = FALSE,
  accept_freesurfer_license = FALSE
)
}
\arguments{
\item{mustWork}{whether to stop if no subjects_dir can be found.}

\item{allow_download}{logical, whether to allow downloading the data in case it is not found locally.}

\item{accept_freesurfer_license}{logical, whether you want to also download fsaverage and fsaverage3, and accept the FreeSurfer license for fsaverage and fsaverage3, available at https://surfer.nmr.mgh.harvard.edu/fswiki/FreeSurferSoftwareLicense. Defaults to FALSE. If FALSE, nothing will be downloaded.}
}
\value{
character string, the subjects directory. If mustWork is FALSE, it will return NULL if no directory was found. If mustWork is TRUE and nothing is found, it stops.
}
\description{
Try a number of ways to find the subjects dir, using environment variables and knowledge on common installation paths of FreeSurfer.
}
\seealso{
Other FreeSurfer helper functions:
\code{\link[=find.freesurferhome]{find.freesurferhome()}},
\code{\link[=fs_home]{fs_home()}},
\code{\link[=has_fs]{has_fs()}}
}
\concept{FreeSurfer helper functions}
