% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{get_surface}
\alias{get_surface}
\title{Extract the surface hemilist from a brainparc.}
\usage{
get_surface(brainparc)
}
\arguments{
\item{brainparc}{a \code{brainparc} instance, see \code{\link{brainparc_fs}} to get one.}
}
\value{
\code{\link{hemilist}} of \code{fs.surface} instances.
}
\description{
Extract the surface hemilist from a brainparc.
}
\seealso{
Other brainparc accessors:
\code{\link[=get_surface_coords]{get_surface_coords()}}
}
\concept{brainparc accessors}
