% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_fs}
\alias{has_fs}
\title{Check whether a FreeSurfer installation can be found on the system.}
\usage{
has_fs()
}
\value{
logical, whether a FreeSurfer installation was found.
}
\description{
Check whether a FreeSurfer installation can be found on the system.
}
\seealso{
Other FreeSurfer helper functions:
\code{\link[=find.freesurferhome]{find.freesurferhome()}},
\code{\link[=fs_home]{fs_home()}},
\code{\link[=get_subjects_dir]{get_subjects_dir()}}
}
\concept{FreeSurfer helper functions}
