% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{is.brainparc}
\alias{is.brainparc}
\title{Check whether object is a brainparc instance.}
\usage{
is.brainparc(x)
}
\arguments{
\item{x}{any `R` object}
}
\value{
logical, \code{TRUE} if parameter \code{x} is a brainparc instance (that is, has "brainparc" in its classes) and \code{FALSE} otherwise.
}
\description{
Check whether object is a brainparc instance.
}
