% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterinfo.R
\name{is.clusterinfo}
\alias{is.clusterinfo}
\title{Check whether x is a clusterinfo instance.}
\usage{
is.clusterinfo(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, whether x is a clusterinfo instance.
}
\description{
Check whether x is a clusterinfo instance.
}
