% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{is.minannot}
\alias{is.minannot}
\title{Check whether object is a minannot instance.}
\usage{
is.minannot(x)
}
\arguments{
\item{x}{any `R` object}
}
\value{
logical, \code{TRUE} if parameter \code{x} is a minannot instance (that is, has "minannot" in its classes) and \code{FALSE} otherwise.
}
\description{
Check whether object is a minannot instance.
}
\keyword{internal}
