% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{mni152reg_mtx}
\alias{mni152reg_mtx}
\title{Get fsaverage (MNI305) to MNI152 transformation matrix.}
\usage{
mni152reg_mtx()
}
\description{
This returns the 4x4 matrix from the FreeSurfer Coordinate Systems documentation to transform between MNI305 and MNI152.
}
\keyword{internal}
