% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_regions.R
\name{name_regions}
\alias{name_regions}
\title{Get names of regions in labeled volume from color lookup table (colorLUT).}
\usage{
name_regions(
  volatlas,
  colorlut,
  ignore_not_in_lut = FALSE,
  warn_not_in_lut = FALSE
)
}
\arguments{
\item{volatlas}{3d integer array, the brain volume segmentation. Can be an \code{fs.volume} instance. Typically something like \code{'subject/mri/aseg.mgz'}.}

\item{colorlut}{a color lookup table data.frame, as returned by \code{\link[freesurferformats]{read.fs.colortable}}. Can be \code{NULL} if you do not have any, in that case the region names will just be 'region_x', where 'x' is the integer region code in the volatlas.}

\item{ignore_not_in_lut}{logical, whether to ignore regions that are not found in the colorLUT. If TRUE, the regions will not appear in the named output list. If FALSE, they will appear as 'region_x', where 'x' is the integer region code in the volatlas.}

\item{warn_not_in_lut}{logical, whether to print a warning if some regions of the volume do not occur in the colorLUT.}
}
\value{
named list, the keys are character strings: the region names from the colorLUT, and the values are integers: the region codes from the volume.
}
\description{
Given a brain volume with integer labels as voxel values and a suitable colorLUT, extract the region names as a named list.
}
