% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_regions.R
\name{named_regions}
\alias{named_regions}
\title{Get names of regions in labeled volume from color lookup table (colorLUT).}
\usage{
named_regions(
  volatlas,
  colorlut,
  ignore_not_in_lut = FALSE,
  warn_not_in_lut = FALSE
)
}
\arguments{
\item{volatlas}{3d integer array, the brain volume segmentation. Can be an \code{fs.volume} instance. Typically something like \code{'subject/mri/aseg.mgz'}.}

\item{colorlut}{a color lookup table data.frame, as returned by \code{\link[freesurferformats]{read.fs.colortable}}.}

\item{ignore_not_in_lut}{logical, whether to ignore regions that are not found in the colorLUT. If TRUE, the regions will not appear in the named output list. If FALSE, they will appear as 'unnamed_region_x', where 'x' is a consecutive integer.}

\item{warn_not_in_lut}{logical, whether to print a warning if some regions of the volume do not occur in the colorLUT.}
}
\value{
named list, the keys are integer region codes, and the values are character strings: the region names from the colorLUT.
}
\description{
Given a brain volume with integer labels as voxel values and a suitable colorLUT, extract the region names as a named list.
}
