% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{print.brainparc}
\alias{print.brainparc}
\title{Print description of a brainparce.}
\usage{
\method{print}{brainparc}(x, ...)
}
\arguments{
\item{x}{brainparc instance, with class `brainparc`.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Called for side effects.
}
\description{
Print description of a brainparce.
}
