% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_atlas_regions.R
\name{region_centroids}
\alias{region_centroids}
\title{Compute centroids of surface atlas parcellation regions.}
\usage{
region_centroids(bp)
}
\arguments{
\item{bp}{a brainparc instance}
}
\value{
hemilist of data.frames, the centroid coordinates for the regions of the respective hemispheres.
}
\description{
Compute centroids of surface atlas parcellation regions.
}
\keyword{internal}
