% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_regions.R
\name{region_color}
\alias{region_color}
\title{Get color string for a region code in a color lookup table.}
\usage{
region_color(colorlut, struct_idx)
}
\arguments{
\item{colorlut}{a color lookup table data.frame, as returned by \code{\link[freesurferformats]{read.fs.colortable}}. Can be \code{NULL} if you do not have any, in that case the region names will just be 'region_x', where 'x' is the integer region code in the volatlas.}

\item{struct_idx}{integer vector, the region / structure codes you are interested in.}
}
\value{
vector of character strings, the hex color names (like '#FF0000' for red).
}
\description{
Get color string for a region code in a color lookup table.
}
\keyword{internal}
