% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{region_colors}
\alias{region_colors}
\title{Get colors for \code{fs.annot} regions.}
\usage{
region_colors(annot, region_names, na_color = "#FFFFFF")
}
\arguments{
\item{annot}{full fs.annot instance}

\item{region_names}{vector of char strings, the query regions for which to get color strings}

\item{na_color}{color to assign to regions which were not found.}
}
\value{
vector of hex color strings
}
\description{
Get colors for \code{fs.annot} regions.
}
\keyword{internal}
