% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loc.R
\name{single_cluster_peaks}
\alias{single_cluster_peaks}
\title{Compute peaks of a single cluster.}
\usage{
single_cluster_peaks(cluster_vertices, statmap, surface, type = "extreme")
}
\arguments{
\item{cluster_vertices}{integer vector, the vertex indices of the cluster.}

\item{statmap}{double vector, the full stat map for the surface. Only the values of the cluster_vertices are used.}

\item{surface}{a single \code{fs.surface} instance. Used to compute the neighborhood of the cluster vertices.}

\item{type}{character string, one of 'extreme', 'min' or 'max'. Which cluster value to report. The default of 'extreme' reports the absolutely larger one of the min and the max value and is for the typical use case of \code{t}-value maps.}
}
\value{
named list with entries 'vertex' and 'value', they contain an integer vector and a double vector, respectively.
}
\description{
Compute peaks of a single cluster.
}
\note{
Called by \code{cluster_peaks}, use that instead.
}
\keyword{internal}
