% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optdata.R
\name{sjd_demo}
\alias{sjd_demo}
\title{Download FreeSurfer template data if needed and return its path.}
\usage{
sjd_demo(accept_freesurfer_license = FALSE)
}
\arguments{
\item{accept_freesurfer_license}{logical, whether you want to also download fsaverage and fsaverage3, and accept the FreeSurfer license for fsaverage and fsaverage3, available at https://surfer.nmr.mgh.harvard.edu/fswiki/FreeSurferSoftwareLicense. Defaults to FALSE. If FALSE, nothing will be downloaded.}
}
\value{
character string, the path to the 'subjects_dir' directory within the downloaded template data directory.
}
\description{
This is a wrapper around \code{download_fsaverage()} and \code{download_fsaverage3())}. It will download the data from the internet unless it already exists locally.
}
\note{
This function will stop if the data cannot be accessed, i.e., the 'subjects_dir' does not exist after trying to download the data.
}
