% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterinfo.R
\name{strvec2int}
\alias{strvec2int}
\title{Convert vector of character strings to integer vector.}
\usage{
strvec2int(input)
}
\arguments{
\item{input}{vector of character strings}
}
\value{
integer vector
}
\description{
Convert vector of character strings to integer vector.
}
\keyword{internal}
