% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{subject_annot_simple_fs}
\alias{subject_annot_simple_fs}
\title{Return simplified annotation: the region names only.}
\usage{
subject_annot_simple_fs(subjects_dir, subject_id, atlas)
}
\arguments{
\item{subjects_dir}{string. The FreeSurfer SUBJECTS_DIR, i.e., a directory containing the data for all your subjects, each in a subdir named after the subject identifier.}

\item{subject_id}{string. The subject identifier}

\item{atlas}{string. The atlas name. E.g., "aparc", "aparc.2009s", or "aparc.DKTatlas". Used to construct the name of the annotation file to be loaded.}
}
\description{
Return simplified annotation: the region names only.
}
\keyword{internal}
