% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subjects_dir}
\alias{subjects_dir}
\title{Find subjects_dir or stop.}
\usage{
subjects_dir()
}
\value{
character string, the subjects directory.
}
\description{
Try a number of ways to find the subjects dir, using environment variables and knowledge on common installation paths of FreeSurfer.
}
