% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusterdata.R
\name{test_clusters_to_annot}
\alias{test_clusters_to_annot}
\title{Internal test function, will be gone soon and converted into separate unit tests.}
\usage{
test_clusters_to_annot(
  sjd = "~/software/freesurfer/subjects",
  sj = "fsaverage"
)
}
\description{
Internal test function, will be gone soon and converted into separate unit tests.
}
\note{
This function is NOT part of the API, using it in client code is a programming error.
}
\keyword{internal}
