% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_atlas_regions.R
\name{vertex_closest_regions}
\alias{vertex_closest_regions}
\title{Find closest regions to vertex using Euclidean or geodesic distance.}
\usage{
vertex_closest_regions(
  brainparc,
  vertices,
  hemis,
  linkage = "single",
  distance = "euclidean"
)
}
\arguments{
\item{brainparc}{a brain parcellation, see functions like \code{\link{brainparc_fs}} to get one.}

\item{vertices}{integer vector, the query vertex indices (from the surface in the \code{brainparc}).}

\item{hemis}{character string vector, the hemispheres for each of the \code{vertices}. Allowed entries are \code{"lh"} and \code{"rh"}, for the left and right brain hemisphere, respectively. Must have same length as the 'vertices' vector (or exactly length 1, in which case we assume that this is the hemi for ALL query vertices).}

\item{linkage}{character string, one of \code{"single"} or \code{"centroid"}. Defines how the distance from a vertex to a region of vertices is computed. \code{"single"}: Euclidean distance from query vertex to the closest vertex of the atlas region. \code{"centroid"}: Euclidean distance from query vertex to the mean of the vertex coordinates of the atlas region.}

\item{distance}{character string, one of \code{"euclidean"} or \code{"geodesic"}. The latter is only supported with \code{linkage = 'single'}.}
}
\description{
Finds the closest atlas regions according to the brain surface parcellation 'brainparc' for the given query vertices.
}
\examples{
\dontrun{
bp = brainparc_fs(fsbrain::fsaverage.path(), "fsaverage", atlas="aparc");
vertex_closest_regions(bp, vertices=c(10, 20), hemis=c("lh", "rh"));
}

}
\seealso{
\code{\link{coord_closest_regions}} if you have a coordinate (on or near the surface) instead of a vertex.
}
