% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_euclid.R
\name{vertex_closest_regions_euclid}
\alias{vertex_closest_regions_euclid}
\title{Find closest regions using Euclidean distance.}
\usage{
vertex_closest_regions_euclid(
  brainparc,
  vertices,
  hemis,
  dist_method = "average"
)
}
\description{
Find closest regions using Euclidean distance.
}
