% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_atlas_regions.R
\name{vertex_regions}
\alias{vertex_regions}
\title{Find regions for vertices in all brainparc atlases.}
\usage{
vertex_regions(brainparc, vertices, hemis)
}
\arguments{
\item{brainparc}{a brain parcellation, see functions like \code{\link{brainparc_fs}} to get one.}

\item{vertices}{integer vector, the query vertex indices (from the surface in the \code{brainparc}).}

\item{hemis}{character string vector, the hemispheres for each of the \code{vertices}. Allowed entries are \code{"lh"} and \code{"rh"}, for the left and right brain hemisphere, respectively. Must have same length as the 'vertices' vector (or exactly length 1, in which case we assume that this is the hemi for ALL query vertices).}
}
\value{
a data.frame, the column names should be obvious.
}
\description{
Find regions for vertices in all brainparc atlases.
}
\examples{
\dontrun{
bp = brainparc_fs(get_subjects_dir(), "fsaverage", atlas="aparc");
vertex_region(bp, vertices=c(10, 20), hemis=c("lh", "rh"));
}

}
\seealso{
\code{\link{vertex_closest_regions}} to compute distance to all regions.
}
