% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{brainparc_fs}
\alias{brainparc_fs}
\title{Create a brain parcellation from a FreeSurfer annotation.}
\usage{
brainparc_fs(subjects_dir, subject_id, surface = "white", atlas = c("aparc"))
}
\arguments{
\item{subjects_dir}{string. The FreeSurfer SUBJECTS_DIR, i.e., a directory containing the data for all your subjects, each in a subdir named after the subject identifier.}

\item{subject_id}{string. The subject identifier}

\item{surface}{string. The surface name. E.g., "white", or "pial". Used to construct the name of the surface file to be loaded.}

\item{atlas}{vector of character strings, the brain atlases to use. For this FreeSurfer version, something like "aparc" for the Desikan-Killiani atlas, "aparc.a2009s" for Destrieux, etc.}
}
\description{
Create a brain parcellation from a FreeSurfer annotation.
}
