% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loc.R
\name{cluster_region_overlap}
\alias{cluster_region_overlap}
\title{Compute overlapping atlas regions for clusters.}
\usage{
cluster_region_overlap(
  clusterinfo,
  silent = getOption("brainloc.silent", default = FALSE)
)
}
\arguments{
\item{clusterinfo}{a clusterinfo instance, see the \code{clusterinfo} function to get one.}

\item{silent}{whether to suppress printing messages to stdout.}
}
\description{
Compute overlapping atlas regions for clusters.
}
