% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusterdata.R
\name{clusterinfo_from_thresholded_overlay}
\alias{clusterinfo_from_thresholded_overlay}
\title{Create clusterinfo instance from thresholded per-vertex data overlays.}
\usage{
clusterinfo_from_thresholded_overlay(
  lh_threshmap,
  rh_threshmap,
  value_thresholded = 0,
  template_subject = "fsaverage",
  subjects_dir = file.path(getOption("brainloc.fs_home", default =
    Sys.getenv("FREESURFER_HOME")), "subjects")
)
}
\arguments{
\item{lh_threshmap}{double vector, the stats map. Typically a thresholded t-value map. Must have one value per vertex. The value assigned to vertices that have been removed by the thresholding can be set with parameter 'value_thresholded'. If a character string, the parameter will be interpreted as file path and loaded with \code{freesurferformats::read.fs.morph}.}

\item{rh_threshmap}{like \code{lh_threshmap}, but for the right hemisphere.}

\item{value_thresholded}{scalar double, the data value of thresholded vertices in the \code{lh_threshmap} and \code{rh_threshmap}.}

\item{template_subject}{character string, the template subject name. Typically 'fsaverage' or 'fsaverage6'. Must be some MNI305 space subject.}

\item{subjects_dir}{character string, file system path to a directory containing the recon-all data for the template_subject. Used to load surfaces and annotations to identify cluster coordinates and atlas regions.}
}
\value{
a \code{clusterinfo} instance
}
\description{
Create clusterinfo instance from thresholded per-vertex data overlays.
}
