% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusterdata.R
\name{clusteroverlay_from_threshmap}
\alias{clusteroverlay_from_threshmap}
\title{Compute separate statsmap and overlay from thresholded map.}
\usage{
clusteroverlay_from_threshmap(threshmap, value_thresholded, surface)
}
\arguments{
\item{value_thresholded}{scalar double, the data value of thresholded vertices in the \code{lh_threshmap} and \code{rh_threshmap}.}

\item{surface}{a single fs.surface instance, used for neighborhood computation.}

\item{theshmap}{double vector, the stats map. Typically a thresholded t-value map. Must have one value per vertex. The value assigned to vertices that have been removed by the thresholding can be set with parameter 'value_thresholded'.}
}
\value{
integer vector, the clusterID overlay
}
\description{
Compute separate statsmap and overlay from thresholded map.
}
\keyword{internal}
