% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_euclid.R
\name{coord_closest_regions_euclid}
\alias{coord_closest_regions_euclid}
\title{Find closest regions to coordinate using Euclidean distance.}
\usage{
coord_closest_regions_euclid(brainparc, coordinate, dist_method = "average")
}
\arguments{
\item{brainparc}{a brain parcellation, see functions like \code{\link{brainparc_fs}} to get one.}

\item{coordinate}{\code{nx3} numerical matrix or vector of length 3, the query point coordinates.}

\item{dist_method}{character string, one of \code{"average"} or \code{"closest"}. Defined how the distance from a vertex to a region of vertices is computed.  \code{"average"}: Euclidean distance from query vertex to the mean of the vertex coordinates of the atlas region. \code{"closest"}: Euclidean distance from query vertex to the closest vertex of the atlas region.}
}
\description{
Find closest regions to coordinate using Euclidean distance.
}
\examples{
\dontrun{
bp = brainparc_fs(fsbrain::fsaverage.path(), "fsaverage", atlas="aparc");
coord_closest_regions_euclid(bp, bp$surfaces$white$lh$vertices[1:3,]);
}

}
