% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coord_MNI152_to_talairach}
\alias{coord_MNI152_to_talairach}
\title{Transform MNI152 coords to Talairach space using Matthew Brett's approach.}
\usage{
coord_MNI152_to_talairach(mni152_coords)
}
\arguments{
\item{mni152_coords}{nx3 numerical matrix of RAS coordinates in MNI152 space.}
}
\value{
nx3 numerical matrix of Talairach coordinates.
}
\description{
See \code{http://brainmap.org/training/BrettTransform.html}.
}
\note{
This is published under the GPL license. See \code{https://github.com/sccn/dipfit/blob/master/mni2tal_matrix.m} and \code{https://github.com/sccn/dipfit/blob/master/mni2tal.m} for a Matlab implementation of the method. All credits go to Matthew Brett.
}
\examples{
    mni_coords = matrix(c(10, 12, 14), nrow = 1, ncol = 3, byrow = TRUE);
    coord_MNI152_to_talairach(mni_coords);

}
