% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_talairach.R
\name{coord_MNI305_to_MNI152}
\alias{coord_MNI305_to_MNI152}
\title{Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.}
\usage{
coord_MNI305_to_MNI152(
  vertex_coords,
  method = getOption("brainloc.method_MNI305_to_from_MNI152", default =
    "best_available")
)
}
\arguments{
\item{vertex_coords}{\code{nx3} numerical matrix of MNI305 surface RAS coordinates, typically from fsaverage surface vertices.}
}
\value{
\code{nx3} numerical matrix of MNI152 coords.
}
\description{
Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates.
}
\note{
One can check that the results are okay by clicking a vertex in FreeView, using the displayed MNI305 coordinates as input to this function, and looking up the reported MNI152 coordinates at \code{https://bioimagesuiteweb.github.io/webapp/mni2tal.html}.
}
\examples{
\dontrun{
# Get MNI152 coordinates for first 3 fsaverage lh vertices:
surf = freesurferformats::read.fs.surface("/opt/freesurfer/subjects/fsaverage/surf/lh.white");
coord_MNI305_to_MNI152(surf$vertices[1:3, ]);
}

}
