% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_subjects_dir}
\alias{get_subjects_dir}
\title{Find subjects_dir or stop.}
\usage{
get_subjects_dir(mustWork = TRUE)
}
\arguments{
\item{mustWork}{whether to stop if no subjects_dir can be found.}
}
\value{
character string, the subjects directory. If mustWork is FALSE, it will return NULL if no directory was found. If mustWork is TRUE and nothing is found, it stops.
}
\description{
Try a number of ways to find the subjects dir, using environment variables and knowledge on common installation paths of FreeSurfer.
}
