% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainparc.R
\name{get_surface_coords}
\alias{get_surface_coords}
\title{Extract the vertex coordinates from the brainparc surface.}
\usage{
get_surface_coords(brainparc, vertices, hemis)
}
\arguments{
\item{brainparc}{a \code{brainparc} instance, see \code{\link{brainparc_fs}} to get one.}

\item{vertices}{integer vector, the query vertex indices.}

\item{hemis}{vector of character string, the hemispheres for the query vertices. Length must match, length of 1 will be recycled to all vertices.}
}
\value{
numerical nx3 matrix, the coordindates for the x query vertices.
}
\description{
Extract the vertex coordinates from the brainparc surface.
}
\seealso{
Other brainparc accessors: 
\code{\link{get_surface}()}
}
\concept{brainparc accessors}
