% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_loc.R
\name{get_talairach_label}
\alias{get_talairach_label}
\title{Retrieve Talairach labels for Talairach coordinates from talairach.org files.}
\usage{
get_talairach_label(
  talfile = "~/develop/talairach/talairach.nii",
  lookup_table_file = "~/develop/talairach/labels.txt",
  tal_coords = matrix(seq.int(15), nrow = 5, ncol = 3)
)
}
\arguments{
\item{talfile}{the NIFTI Talairach label file from talairach.org}

\item{lookup_table_file}{the Talairach label index text file from talairach.org}

\item{tal_coords}{nx3 numeric matrix, the n query Talairach coordinates}
}
\value{
the labels for the coordinates
}
\description{
Retrieve Talairach labels for Talairach coordinates from talairach.org files.
}
\note{
You need to download the required files from \code{talairach.org}.
}
\keyword{internal}
