% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hemilist}
\alias{hemilist}
\title{Create hemilist.}
\usage{
hemilist(lh_data = NULL, rh_data = NULL)
}
\arguments{
\item{lh_data}{any R object, the lh value.}

\item{rh_data}{any R object, the rh value.}
}
\value{
a hemilist
}
\description{
A hemilist is a named list which has at least one of the following fields: \code{lh} and \code{rh}. The fields can store arbitrary values.
}
\note{
This function is trivial and mainly exists to document what a hemilist is (so this explanation can be linked to from the documentation of other functions which use the concept).
}
\seealso{
Other hemilist functions: 
\code{\link{hemilist.wrap}()}
}
\concept{hemilist functions}
