% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.hemilist}
\alias{is.hemilist}
\title{Find out whether x is a hemilist}
\usage{
is.hemilist(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, whether x is a hemilist
}
\description{
Find out whether x is a hemilist
}
\keyword{internal}
