% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_annot.R
\name{num_clusters}
\alias{num_clusters}
\title{Compute number of clusters from cluster annot or clusterinfo instance.}
\usage{
num_clusters(cluster_annots)
}
\arguments{
\item{cluster_annots}{\code{\link{hemilist}} of cluster annots, see \code{\link{clusteroverlay_to_annot}}.}
}
\value{
named list with keys 'lh', 'rh' and 'total', each holding a scalar integer. The cluster counts.
}
\description{
Compute number of clusters from cluster annot or clusterinfo instance.
}
\keyword{internal}
