% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_regions.R
\name{segmentation_centers}
\alias{segmentation_centers}
\title{Find center of mass of regions in a brain volume segmentation.}
\usage{
segmentation_centers(volatlas, named_regions = NULL, vox2ras = diag(4))
}
\arguments{
\item{volatlas}{3d integer array, the brain volume segmentation. Can be an \code{fs.volume} instance. Typically something like \code{'subject/mri/aseg.mgz'}.}

\item{named_regions}{the regions to consider, a named list where the keys are character strings: the region names from the colorLUT, and the values are integers: the region codes from the volume. See \code{\link{name_regions}} to get one. Can be \code{NULL}, in which case all regions in the volatlas will be used, and the region names will appear as 'region_x', where 'x' is the integer region code in the volatlas.}

\item{regions}{}
}
\value{
data.frame with region names and the center of mass x, y and z coordinates for the regions.
}
\description{
Find center of mass of regions in a brain volume segmentation.
}
\examples{
\dontrun{
fsh = brainloc:::fs.home();
lutfile = file.path(fsh, "FreeSurferColorLUT.txt");
segfile = file.path(fsh, 'subjects', 'fsaverage', 'mri', 'aseg.mgz');
named_regions = name_regions(segfile, lutfile);
named_regions$Unknown = NULL; # Ignore an atlas region.
sc = segmentation_centers(segfile, named_regions);
regions_distance_matrix = dist(sc);
}

}
